refactor(dedicated): fix comments
 ref: MANAGER-12284
