+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/dashboard/installation/choice/index.js
@@ -6,29 +6,22 @@ import 'angular-translate';
 
 const moduleName = 'ovhManagerDedicatedServerDashboardServerInstallationChoice';
 
-angular
-  .module(moduleName, [
-    'ui.router',
-    'oc.lazyLoad',
-    'ngTranslateAsyncLoader',
-    'pascalprecht.translate',
-  ])
-  .config(
-    /* @ngInject */ ($stateProvider) => {
-      $stateProvider.state(
-        'app.dedicated-server.server.dashboard.installation-choice.**',
-        {
-          url: '/installation/choice',
-          lazyLoad: ($transition$) => {
-            const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
+angular.module(moduleName, ['ui.router', 'oc.lazyLoad']).config(
+  /* @ngInject */ ($stateProvider) => {
+    $stateProvider.state(
+      'app.dedicated-server.server.dashboard.installation-choice.**',
+      {
+        url: '/installation/choice',
+        lazyLoad: ($transition$) => {
+          const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
 
-            return import('./choice.module').then((mod) =>
-              $ocLazyLoad.inject(mod.default || mod),
-            );
-          },
+          return import('./choice.module').then((mod) =>
+            $ocLazyLoad.inject(mod.default || mod),
+          );
         },
-      );
-    },
-  );
+      },
+    );
+  },
+);
 
 export default moduleName;
