+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/dashboard/installation/gabarit/gabarit.module.js
@@ -5,9 +5,6 @@ import routing from './gabarit.routing';
 const moduleName =
   'ovhManagerDedicatedServerDashboardServerInstallationGabarit';
 
-angular
-  .module(moduleName, ['ui.router'])
-  .config(routing)
-  .run(/* @ngTranslationsInject ./translations */);
+angular.module(moduleName, ['ui.router']).config(routing);
 
 export default moduleName;
