+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/dashboard/installation/ovh/ovh.module.js
@@ -4,9 +4,6 @@ import routing from './ovh.routing';
 
 const moduleName = 'ovhManagerDedicatedServerDashboardServerInstallationOvh';
 
-angular
-  .module(moduleName, ['ui.router'])
-  .config(routing)
-  .run(/* @ngTranslationsInject ./translations */);
+angular.module(moduleName, ['ui.router']).config(routing);
 
 export default moduleName;
