+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/dashboard/installation/progress/index.js
@@ -7,29 +7,22 @@ import 'angular-translate';
 const moduleName =
   'ovhManagerDedicatedServerDashboardServerInstallationProgress';
 
-angular
-  .module(moduleName, [
-    'ui.router',
-    'oc.lazyLoad',
-    'ngTranslateAsyncLoader',
-    'pascalprecht.translate',
-  ])
-  .config(
-    /* @ngInject */ ($stateProvider) => {
-      $stateProvider.state(
-        'app.dedicated-server.server.dashboard.installation-progress.**',
-        {
-          url: '/installation/progress',
-          lazyLoad: ($transition$) => {
-            const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
+angular.module(moduleName, ['ui.router', 'oc.lazyLoad']).config(
+  /* @ngInject */ ($stateProvider) => {
+    $stateProvider.state(
+      'app.dedicated-server.server.dashboard.installation-progress.**',
+      {
+        url: '/installation/progress',
+        lazyLoad: ($transition$) => {
+          const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
 
-            return import('./progress.module').then((mod) =>
-              $ocLazyLoad.inject(mod.default || mod),
-            );
-          },
+          return import('./progress.module').then((mod) =>
+            $ocLazyLoad.inject(mod.default || mod),
+          );
         },
-      );
-    },
-  );
+      },
+    );
+  },
+);
 
 export default moduleName;
