+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/dashboard/installation/progress/progress.module.js
@@ -5,9 +5,6 @@ import routing from './progress.routing';
 const moduleName =
   'ovhManagerDedicatedServerDashboardServerInstallationProgress';
 
-angular
-  .module(moduleName, ['ui.router'])
-  .config(routing)
-  .run(/* @ngTranslationsInject ./translations */);
+angular.module(moduleName, ['ui.router']).config(routing);
 
 export default moduleName;
