+++ packages/manager/modules/bm-server-components/src/general-information/install/image/components/config/component.js
@@ -1,5 +1,5 @@
 import controller from './controller';
-import template from './index.html';
+import template from './template.html';
 
 export default {
   name: 'serverInstallImageConfig',
