+++ packages/manager/modules/bm-server-components/src/general-information/install/image/components/config/template.html
