+++ packages/manager/modules/bm-server-components/src/general-information/install/image/template.html
